+++ src/Client/RestAPI/DiscordRejection.ts
