+++ src/Interfaces/Interfaces.ts
@@ -9,6 +9,7 @@ import { UsersManager } from "../Client/Managers/UsersManager";
 import { Overwrite } from "../Structures/Channel/Overwrite";
 import { CHANNELTYPES } from "../Utils/Constants";
 import { MessageEmbed } from "../Utils/Embed/MessageEmbed";
+import { ICreateGuildIntegration, IGuildIntegration } from "./Integration";
 
 export interface CacheProviders {
   guilds?: GuildsManager;
@@ -21,10 +22,13 @@ export interface CacheProviders {
 
 export interface IAPIParams {
   endpoint: string;
-  method: "GET" | "POST" | "DELETE" | "PUT";
-  postType?: "Message" | "Channel";
+  method: "GET" | "POST" | "DELETE" | "PUT" | "PATCH";
+  postType?: "Message" | "Channel" | "Integration" | "[Message]" | "JSON";
   message?: MessageOptions;
   channel?: ChannelOptions;
+  integration?: ICreateGuildIntegration;
+  messages?: string[];
+  json_params?: Object;
 }
 
 export interface MessageOptions {
@@ -46,9 +50,9 @@ export interface ChannelOptions {
   nsfw?: boolean;
 }
 
-export interface Payload {
+export interface Payload<D = any> {
   op: OPCODE;
   t?: string;
   s?: number;
-  d?: any;
+  d?: D;
 }
