+++ src/index.ts
@@ -4,8 +4,8 @@ export * from "./Client/ClientOptions";
 export * from "./Decorators/Builder";
 export * from "./Decorators/Events";
 export * from "./Client/ClientUser";
-export * from "./Client/API/RestAPI";
-export * from "./Client/API/RestAPIHandler";
+export * from "./Client/RestAPI/RestAPI";
+export * from "./Client/RestAPI/RestAPIHandler";
 export * from "./Client/Websocket/Gateway";
 export * from "./Client/Websocket/Websocket";
 export * from "./Client/Events/BaseEvent";
